% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colar.R
\name{colar}
\alias{colar}
\title{Função Colar}
\usage{
colar(a = NULL, b = NULL)
}
\arguments{
\item{a}{primeiro parâmetro da função}

\item{b}{segund parâmetro da função}
}
\value{
retorna o resultado do Paste das duas variáveis recebidas
}
\description{
Esta função faz o Paste/Colar de duas variáveis recebidas
}
